Shop now supports buying and selling heroes
Temporary buttons are used while drag and drop is being built. You must have open party slots in order to buy heroes. You will encounter an error if you try to buy or sell from a slot with nothing in it, this should be a non issue when drag and drop is implemented.